tests: Port shared SQL container + per-class DB restore to xUnit fixtures
Reshapes the MSTest SharedSqlServer/AzureSqlDbContainer pair into the
xUnit fixture model while preserving the perf-critical "boot the
container once per assembly, RESTORE a per-class DB from a schema-only
template" behaviour:

- SharedSqlFixture (IAsyncLifetime, used via ICollectionFixture):
  boots the MsSqlContainer once for the assembly, prepares the
  AccountsTemplate backup, disposes the container on assembly
  teardown. Replaces SharedSqlServer + its [AssemblyCleanup] hook.
- PerClassDbFixture (IAsyncLifetime, used via IClassFixture): on first
  use by a class, RESTOREs a fresh per-class DB from the template;
  drops it on class teardown (SINGLE_USER WITH ROLLBACK IMMEDIATE
  first, to evict lingering connections). InitializeAsync /
  DisposeAsync are virtual so AuditLogs* subclasses can layer
  per-class seeding on top.

The 3.5 GiB container memory cap gains an inline comment linking it
to maxParallelThreads = 4 in xunit.runner.json -- they need to move
together to avoid OOM.
